Remove obsolete "known problem" from Windows instructions
Also clarify a bit that it's not interfacing with other Windows
libraries that is possible with the MSVC toolchain, but only with ones
that are incompatible with MINGW. Which are not that many after all.

And add a comment that using the MSVC toolchain allows using the Visual
Studio debugger.